// lib/types.ts
export interface Product {
  id: string
  name: string
  description: string
  category: string
  brand: string
  price: number
  stock: number
  rating: number
  available: boolean
  color: string
  weightKg: string
  dimensionsCm: string
  warrantyMonths: number
  material: string
  createdAt: string
  updatedAt: string
}